+++ src/main/java/com/jcloisterzone/action/TunnelAction.java
@@ -1,8 +1,10 @@
 package com.jcloisterzone.action;
 
 import com.jcloisterzone.board.pointer.FeaturePointer;
+import com.jcloisterzone.game.Token;
 import com.jcloisterzone.ui.GameController;
 import com.jcloisterzone.ui.annotations.LinkedImage;
+import com.jcloisterzone.wsio.message.PlaceTokenMessage;
 
 import io.vavr.collection.Set;
 
@@ -10,21 +12,21 @@ import io.vavr.collection.Set;
 @LinkedImage("actions/tunnel")
 public class TunnelAction extends SelectFeatureAction {
 
-    private final boolean secondTunnelPiece;
+    private final Token token;
 
-    public TunnelAction(Set<FeaturePointer> ptrs, boolean secondTunnelPiece) {
-        super(ptrs);
-        this.secondTunnelPiece = secondTunnelPiece;
+    public TunnelAction(Set<FeaturePointer> options, Token token) {
+        super(options);
+        assert token.isTunnel();
+        this.token = token;
     }
 
-    public boolean isSecondTunnelPiece() {
-        return secondTunnelPiece;
+    public Token getToken() {
+        return token;
     }
 
     @Override
-    public void perform(GameController gc, FeaturePointer bp) {
-        //server.placeTunnelPiece(bp, secondTunnelPiece);
-        //TODO
+    public void perform(GameController gc, FeaturePointer ptr) {
+        gc.getConnection().send(new PlaceTokenMessage(gc.getGameId(), token, ptr));
     }
 
     @Override