+++ src/main/java/com/jcloisterzone/game/Token.java
@@ -1,5 +1,7 @@
 package com.jcloisterzone.game;
 
+import io.vavr.collection.List;
+
 public enum Token {
 
     WINE, GRAIN, CLOTH,
@@ -11,5 +13,13 @@ public enum Token {
     CASTLE,
     FERRY,
     GOLD,
-    TUNNEL_A, TUNNEL_B, TUNNEL_C
+    TUNNEL_A, TUNNEL_B, TUNNEL_C;
+
+    public boolean isTunnel() {
+        return this == TUNNEL_A || this == TUNNEL_B || this == TUNNEL_C;
+    }
+
+    public static Token[] tunnelValues() {
+        return new Token[] {Token.TUNNEL_A, Token.TUNNEL_B, Token.TUNNEL_C};
+    }
 }