+++ src/main/java/com/jcloisterzone/ui/ImmutablePoint.java
@@ -1,5 +1,8 @@
 package com.jcloisterzone.ui;
 
+import java.awt.Point;
+import java.awt.geom.AffineTransform;
+
 import com.jcloisterzone.Immutable;
 import com.jcloisterzone.board.Rotation;
 
@@ -26,6 +29,13 @@ public class ImmutablePoint {
 
     public ImmutablePoint add(int tx, int ty) {
         return new ImmutablePoint(x + tx, y + ty);
+    }
+
+    public ImmutablePoint transform(AffineTransform tx) {
+        Point ptSrc = new Point(x, y);
+        Point ptDst = new Point();
+        tx.transform(ptSrc, ptDst);
+        return new ImmutablePoint(ptDst.x, ptDst.y);
     }
 
     public ImmutablePoint scale(int tileWidth, int tileHeight) {