+++ src/main/java/com/jcloisterzone/ui/view/GameSetupView.java
@@ -23,6 +23,7 @@ import com.jcloisterzone.ui.Client;
 import com.jcloisterzone.ui.GameController;
 import com.jcloisterzone.ui.MenuBar;
 import com.jcloisterzone.ui.MenuBar.MenuItem;
+import com.jcloisterzone.ui.PlayerColor;
 import com.jcloisterzone.ui.controls.chat.ChatPanel;
 import com.jcloisterzone.ui.controls.chat.GameChatPanel;
 import com.jcloisterzone.ui.panel.BackgroundPanel;
@@ -139,20 +140,26 @@ public class GameSetupView extends AbstractUiView {
     @Subscribe
     public void onGameStarted(GameStartedEvent ev) {
         Stream<PlayerSlot> slots = Stream.ofAll(Arrays.asList(gc.getGame().getPlayerSlots()));
-        Array<PlayerSlot> occupiedSlots = slots.filter(slot -> slot.isOccupied()).toArray();
-        Array<PlayerSlot> freeSlots = slots.filter(slot -> !slot.isOccupied()).toArray();
+        Array<PlayerSlot> occupiedSlots = slots.filter(slot -> slot != null && slot.isOccupied()).toArray();
+        // for free color we can't search slot - because for load game, slots are already filtered
+        // to existing ones
+        Array<PlayerColor> freeColors = Stream.range(0, PlayerSlot.COUNT)
+            .filter(i -> occupiedSlots.find(s -> s.getNumber() == i).isEmpty())
+            .map(i -> gc.getConfig().getPlayerColor(i))
+            .toArray();
+
         int occupiedSize = occupiedSlots.size();
-        int freeSize = freeSlots.size();
+        int freeSize = freeColors.size();
         int i = 0;
         for (PlayerSlot slot : occupiedSlots) {
             Map<Token, Color> tunnelColors = new HashMap<>();
             tunnelColors.put(Token.TUNNEL_A, slot.getColors().getMeepleColor());
             if (freeSize >= occupiedSize) {
-                tunnelColors.put(Token.TUNNEL_B, freeSlots.get(i).getColors().getMeepleColor());
+                tunnelColors.put(Token.TUNNEL_B, freeColors.get(i).getMeepleColor());
                 i++;
             }
             if (freeSize >= 2 * occupiedSize) {
-                tunnelColors.put(Token.TUNNEL_C, freeSlots.get(i).getColors().getMeepleColor());
+                tunnelColors.put(Token.TUNNEL_C, freeColors.get(i).getMeepleColor());
                 i++;
             }
             slot.getColors().setTunnelColors(tunnelColors);